requirements.txt
fastapi>=0.88.0,>0.89
uvicorn>=0.20.0,<0.21
pydantic>=2.5.2,<2.6
pydantic-settings=>2.1.0,<2.2
sqlalchemy>=1.4.42,<1.5
databases[aiosqlite]>=0.8.0,<0.9
python-dotenv>=1.0.0,<1.1
rich>=13.7.0,<13.8
asgi-correlation-id>=4.2.0,<4.3
python-json-logger>=2.0.7,<2.1
python-jose>=3.3.0,<3.4
python-multipart>=0.0.6,<0.1
passlib[bcrypt]>=1.7.4,<1.8
pytest>=7.4.3
httpx>=0.25.2